The shell holds its structural limits, and the bar chrome stands alone
The lane's growth broke three ADR-0015 structural limits. Two split a
whole concern out; one moves a grandfather pin deliberately:

- The control bar's standalone faces — the arm telegraph, connection
  chip, unavailable-tile placard, and session log sheet — move to
  ControlBarChrome.swift. Each reads the model's published surface
  only, so the move loosens nothing.
- The link driver's three reader tasks move to driver/readers.rs:
  bytes-to-events pumps with no engine knowledge of their own.
- HostLinkModel.swift re-pins at its exact size. Its relay targets,
  controller loop, and published state are one main-actor weave —
  Swift's file-scoped `private` means splitting it needs a real
  sub-object design, and pretending an extension-file move is that
  design would trade encapsulation for a line count. The pin keeps
  the file shrink-only until that design happens.
- The formatter's reflow of two link sources rides along, so the
  facade's fmt gate holds under the CI toolchain's rustfmt.
